+++ atplug-plugin-gradle/src/main/java/com/diffplug/atplug/tooling/PlugGeneratorJavaExecable.kt
@@ -1,0 +1,40 @@
+/*
+ * Copyright (C) 2022 DiffPlug
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.diffplug.atplug.tooling
+
+import com.diffplug.atplug.tooling.gradle.JavaExecable
+import java.io.File
+import java.util.*
+
+/** [PlugGenerator.PlugGenerator] in a [JavaExecable] form. */
+class PlugGeneratorJavaExecable(toSearch: List<File>?, toLinkAgainst: Set<File>?) : JavaExecable {
+	// inputs
+	var toSearch: List<File>
+	var toLinkAgainst: Set<File>
+
+	// outputs
+	@JvmField var atplugInf: SortedMap<String, String>? = null
+
+	init {
+		this.toSearch = ArrayList(toSearch)
+		this.toLinkAgainst = LinkedHashSet(toLinkAgainst)
+	}
+
+	override fun run() {
+		val metadataGen = PlugGenerator(toSearch, toLinkAgainst)
+		atplugInf = metadataGen.atplugInf
+	}
+}